⚡️:: 내 학교 검증

+++ src/main/java/com/walkhub/walkhub/domain/challenge/service/RemoveChallengeService.java
@@ -2,6 +2,7 @@ package com.walkhub.walkhub.domain.challenge.service;
 
 import com.walkhub.walkhub.domain.challenge.domain.Challenge;
 import com.walkhub.walkhub.domain.challenge.domain.repository.ChallengeRepository;
+import com.walkhub.walkhub.domain.challenge.exception.NotYourSchoolException;
 import com.walkhub.walkhub.domain.challenge.facade.ChallengeFacade;
 import com.walkhub.walkhub.domain.user.domain.User;
 import com.walkhub.walkhub.domain.user.facade.UserFacade;
@@ -28,6 +29,11 @@ public class RemoveChallengeService {
             throw InvalidRoleException.EXCEPTION;
         }
 
+        if (!challenge.isMySchool(user.getRealSchoolAgencyCode())){
+            throw NotYourSchoolException.EXCEPTION;
+        }
+
         challengeRepository.delete(challenge);
     }
+
 }
